implemented trading assets between organisational units after creating buy offer

+++ src/ElectronicAssetTradingPlatform/AssetTrading/BuyOffer.java
@@ -50,7 +50,25 @@ public class BuyOffer extends Offer{
         return BuyOffersDB.getBuyOffersDB().getMarketBuyOffers().lastKey() + 1;
     }
 
-    @Override
+
+    // overriden, function which also exchanges assets between buyer and seller
+    public void resolveOffer(OrganisationalUnit buyer, OrganisationalUnit seller) throws Exception {
+        // loop until there is no matching offer OR this.quantity == 0
+        while(checkMatchedOffer() != 0 && this.getQuantity() > 0) {
+            int matchingID = checkMatchedOffer();
+
+            // reduce the quantities of matching buy and sell offers + deleting offers if they've been fully resolved
+            reduceOrderQuantities(matchingID, buyer, seller);
+
+            // buy offer is fully resolved
+            if (this.getQuantity() <= 0) {
+                long millis = System.currentTimeMillis();
+                this.dateResolved = new Date(millis);
+            }
+        }
+    }
+
+    // only updates the offers, doesn't actually trade assets - only here for testing functionality delete later
     public void resolveOffer() {
         // loop until there is no matching offer OR this.quantity == 0
         while(checkMatchedOffer() != 0 && this.getQuantity() > 0) {
@@ -80,8 +98,8 @@ public class BuyOffer extends Offer{
             // if the quantity of buy offer is greater than the sell offer, remove the sell offer from DB
             // and reduce the quantity of the buy offer
             else if (this.getQuantity() > matchingSellOffer.getQuantity()) {
-                this.setQuantity(this.getQuantity() - matchingSellOffer.getQuantity());
-                // update the database with new quantity
+                int quantityTraded = matchingSellOffer.getQuantity();
+                this.setQuantity(this.getQuantity() - quantityTraded);
                 BuyOffersDB.addBuyOffer(this.getOfferID(), this);
                 matchingSellOffer.setQuantity(0);
                 SellOffersDB.removeSellOffer(matchingID);
@@ -89,6 +107,57 @@ public class BuyOffer extends Offer{
             // if the quantity of buy offers is less than the sell offers, remove the buy offer from DB
             // and reduce the quantity of the sell offer
             else {
+                matchingSellOffer.setQuantity(matchingSellOffer.getQuantity() - this.getQuantity());
+                SellOffersDB.addSellOffer(matchingID, matchingSellOffer);
+                BuyOffersDB.removeBuyOffer(this.getOfferID());
+                this.setQuantity(0);
+            }
+        }
+    }
+
+
+    // overridden, function which also exchanges assets between buyer and seller
+    private void reduceOrderQuantities(int matchingID, OrganisationalUnit buyOrg, OrganisationalUnit sellOrg) throws Exception {
+        if (matchingID != 0) {
+            SellOffer matchingSellOffer = SellOffersDB.getSellOffersDB().getOffer(matchingID);
+            double sellersPrice = matchingSellOffer.getPricePerUnit();
+            // if the quantity of buy and sell offers are equal remove them both from the DB
+            if (this.getQuantity() == matchingSellOffer.getQuantity()) {
+                int quantityTraded = this.getQuantity();
+                buyOrg.addAsset(this.getAssetName(), quantityTraded);
+                sellOrg.removeAsset(this.getAssetName(), quantityTraded);
+                sellOrg.editCredits(sellersPrice * (double)quantityTraded);
+                buyOrg.editCredits(-(sellersPrice * (double)quantityTraded));
+
+
+                SellOffersDB.removeSellOffer(matchingID);
+                this.setQuantity(0);
+                BuyOffersDB.removeBuyOffer(this.getOfferID());
+                matchingSellOffer.setQuantity(0);
+            }
+            // if the quantity of buy offer is greater than the sell offer, remove the sell offer from DB
+            // and reduce the quantity of the buy offer
+            else if (this.getQuantity() > matchingSellOffer.getQuantity()) {
+                int quantityTraded = matchingSellOffer.getQuantity();
+                buyOrg.addAsset(this.getAssetName(), quantityTraded);
+                sellOrg.removeAsset(this.getAssetName(), quantityTraded);
+                sellOrg.editCredits(sellersPrice * (double)quantityTraded);
+                buyOrg.editCredits(-(sellersPrice * (double)quantityTraded));
+
+                this.setQuantity(this.getQuantity() - quantityTraded);
+                BuyOffersDB.addBuyOffer(this.getOfferID(), this);
+                matchingSellOffer.setQuantity(0);
+                SellOffersDB.removeSellOffer(matchingID);
+            }
+            // if the quantity of buy offers is less than the sell offers, remove the buy offer from DB
+            // and reduce the quantity of the sell offer
+            else {
+                int quantityTraded = this.getQuantity();
+                buyOrg.addAsset(this.getAssetName(), quantityTraded);
+                sellOrg.removeAsset(this.getAssetName(), quantityTraded);
+                sellOrg.editCredits(sellersPrice * (double)quantityTraded);
+                buyOrg.editCredits(-(sellersPrice * (double)quantityTraded));
+
                 matchingSellOffer.setQuantity(matchingSellOffer.getQuantity() - this.getQuantity());
                 SellOffersDB.addSellOffer(matchingID, matchingSellOffer);
                 BuyOffersDB.removeBuyOffer(this.getOfferID());
